+++ Packages/com.unity.multiplayer.samples.coop/Utilities/SceneManagement/SceneLoaderWrapper.cs
@@ -154,7 +154,6 @@ namespace Unity.Multiplayer.Samples.Utilities
                     if (NetworkManager.IsClient)
                     {
                         m_ClientLoadingScreen.StopLoadingScreen();
-                        m_LoadingProgressManager.ResetLocalProgress();
                     }
                     break;
                 case SceneEventType.Synchronize: // Server told client to start synchronizing scenes