+++ Packages/com.unity.multiplayer.samples.coop/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 * Clarified fix used to handle clients reconnecting to the server while keeping the same active scene and having additional scenes additively loaded (#864)
+* Host loading progress bar is now properly updated on all clients during scene loads (#862)
 
 ## [1.6.1] - 2023-06-14